refactor: admin module restructurisation
