+++ src/app/modules/admin/admin.module.ts
@@ -6,14 +6,14 @@ import { RouterModule, Routes } from '@angular/router';
 import { UserConfigResolver } from 'src/app/core/resolvers/user-config.resolver';
 import { SharedModule } from 'src/app/shared/shared.module';
 import { MaterialModule } from '../material/material.module';
-import { AdminHomeComponent } from './views/admin-home/admin-home.component';
+import { AdminComponent } from './admin.component';
 import { TutorDetailedViewComponent } from './views/tutor-detailed-view/tutor-detailed-view.component';
 import { TutorsListViewComponent } from './views/tutors-list-view/tutors-list-view.component';
 
 const routes: Routes = [
   {
     path: '',
-    component: AdminHomeComponent,
+    component: AdminComponent,
     children: [
       { path: 'tutors', component: TutorsListViewComponent, resolve: { config: UserConfigResolver } },
       { path: 'tutor/:id', component: TutorDetailedViewComponent },
@@ -22,7 +22,7 @@ const routes: Routes = [
 ];
 
 @NgModule({
-  declarations: [AdminHomeComponent, TutorsListViewComponent, TutorDetailedViewComponent],
+  declarations: [TutorsListViewComponent, TutorDetailedViewComponent, AdminComponent],
   imports: [CommonModule, MaterialModule, SharedModule, FlexLayoutModule, FormsModule, ReactiveFormsModule, RouterModule.forChild(routes)],
 })
 export class AdminModule {}
